requirements.txt
langchain
langchain-community
langchain-groq
unstructured
selenium
streamlit
pandas
python-dotenv
langchain-huggingface
faiss-cpu